Regression(PSON): ViewGestureController is not properly notified of process swaps on iOS https://bugs.webkit.org/show_bug.cgi?id=196029 <rdar://problem/48954651>
Reviewed by Tim Horton.

Source/WebKit:

If there is a ViewGestureController when process swapping, make sure we disconnect it
from the old process and reconnect it to the new one. This matches what is done in
WebViewImpl for macOS (see r238356).

* UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _processWillSwap]):
(-[WKWebView _processDidExit]):
(-[WKWebView _didRelaunchProcess]):

Tools:

Add API test coverage.

* TestWebKitAPI/Tests/WebKitCocoa/ProcessSwapOnNavigation.mm:


Identifier: 210306@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@243240 268f45cc-cd09-0410-ab3c-d52691b4dbfc